+++ lambdachecker-easy-submit/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
-All notable changes to the "lambdachecker-easy-submit" extension will be documented in this file.
+All notable changes to the "lambdachecker" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
 - Initial release
+
+Available features:
+- Unlock a contest by clicking on it in the Treeview and providing the password
+- Locked/Unlocked status for contests is shown with a FileDecoration